+++ experimental/packages/opentelemetry-configuration/src/FileConfigProvider.ts
@@ -16,6 +16,7 @@
 
 import { diagLogLevelFromString, getStringFromEnv } from '@opentelemetry/core';
 import {
+  ConfigAttributes,
   ConfigurationModel,
   initializeDefaultConfiguration,
 } from './configModel';
@@ -104,8 +105,7 @@ function parseConfigFile(config: ConfigurationModel) {
   }
 }
 
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-function setResourceAttributes(config: ConfigurationModel, attributes: any[]) {
+function setResourceAttributes(config: ConfigurationModel, attributes: ConfigAttributes[]) {
   if (attributes) {
     config.resource.attributes = [];
     for (let i = 0; i < attributes.length; i++) {
@@ -135,8 +135,13 @@ function setResourceAttributes(config: ConfigurationModel, attributes: any[]) {
       }
 
       config.resource.attributes.push({
+<<<<<<< HEAD
         name: att['name'],
         value: value,
+=======
+        name: getStringFromConfigFile(att['name']) ?? '',
+        value: att['value'],
+>>>>>>> 36051e88440a9f78d07bdc0a605e4e79a4ce1cde
         type: att['type'] ?? 'string',
       });
     }